Modularise backends.
Move the code specific to Docker and Singularity into distinct "backend
modules". The idea is that most of the runner code should be independent
of the backend used, and all the code specific to one particular backend
should fit in one module clearly separated from the rest of the code.