Automatically set Java memory options.
We add an `info` member struct to the `odk_backend` struct, intended to
hold informations obtained from the backend. For now, this struct has
only one member, `total_memory`, for the total amount of memory
available to the backend.

We update the initialisation function of each backend to fill the struct
when the backend is initialised. For the native and singularity
backends, the total amount of memory available to the backend is simply
the total physical memory of the underlying system. For the Docker
backend, we obtain the max amount of memory that Docker is configured to
use by calling `docker info` (this will return the physical amount of
memory if Docker has no configured memory limits).

After the backend has been initialised, we use the information about the
total memory to automatically set the Java memory options so that Java
processes will be allowed to use up to 90% of the available memory.