Allow running as root withing a container.
The ODK image expects to get passed a ODK_USER_ID environment variable
(along with a ODK_GROUP_ID) indicating the UID of the calling user. On
GNU/Linux, this is using to ensure that all processes within the
container run under that UID, therefore ensuring that any file created
ultimately belongs to the calling user (rather than to root, under whose
identity the Docker deamon typically runs).

If no ODK_USER_ID variable is set, the ODK image defaults to run
processes under a UID of 1000, which may or may not be suitable (1000 is
the UID of the first real user account on most, if not all, GNU/Linux
distributions).

If ODK_USER_ID is set to zero, processes inside the container run under
the identity of the superuser.

This commit makes sure that the ODK_USER_ID / ODK_GROUP_ID variables are
passed to the Docker or Singularity backend. They default to the real
UID and GID of the caller, unless the `--root` option is used, in which
case they are set to zero.

This requires the addition of a new backend method, `prepare`, which is
to be called prior to `run` to give the backend an opportunity to update
the runner's configuration (in this case, to add the required
environment variables).