Store Java options and properties separately.
For now, we have only one Java option that we may need to pass to the
container (-Xmx, to specify the max amount of memory). But in the
future, we may need to pass more than that, for example:

* options to allow the use of a proxy;
* properties to configure the OWLAPI;
* arbitrary properties the user may wish to pass.

It would not be practical to store all those options directly into the
ODK_JAVA_OPTS environment variable, because we would then need to update
that variable as many times as we have options to pass.

Instead, we will store the Java options and properties in a separate
array in the ODK configuration object. We reuse the odk_env_var_t
structure, which we rename into odk_var_t since it is no longer used
only for environment variables.

When we are ready to invoke a container, we convert the contents of that
array into the ODK_JAVA_OPTS and ROBOT_JAVA_ARGS environment variables
that ODK and ROBOT expect.