Detect if we are in an ODK-managed repository.
Do not systematically set the working directory inside the container to
/work/src/ontology and do not systematically bind $PWD/../.. to /work.

We should only do that if we are in the src/ontology directory of an
ODK-managed repository; otherwise, the working directory should simply
be /work, which should be bound to the caller's current directory.

We detect an ODK-managed repository by (1) looking for a "*-odk.yaml"
file in the current directory, and (2) checking that the pathname of the
current directory ends with "/src/ontology".