Forward SSH auth socket to the container.
When using a container, if we detect a SSH authentication socket
advertised in the environment (as SSH_AUTH_SOCK), we forward it to the
Docker container by (1) binding the path to the socket on the host side
to a known path on the container side and (2) setting the SSH_AUTH_SOCK
environment variable in the container to point to said known path.

Not tested on Windows yet, not sure if and how that will work there.