Allow binding an inexistent path on the host.
The realpath function returns NULL if the path it is given does not
exist (among other error conditions). However there is at least one
legitimate reason for creating a binding to a non-existent path: to bind
the special socket /run/host-services/ssh-auth.sock on macOs. That
socket does not actually exist, but Docker on macOS recognises its name
when it is used in a binding and uses it to enable forwarding the SSH
authentication socket to the container.

So instead of directly using the value returned by realpath (which might
be NULL), we check if realpath failed because the specified path does
not exist, and if so, we just use the provided path as is.