Force "posixly correct" behaviour of getopt.
When parsing the command line, force `getopt` to use its "posixly
correct" mode, in which arguments starting with `-` are no longer
recognised as options if they occur after any positional argument.

This is for convenience, as this allows using for example a command like
the following:

  $ odkrun make -B

Without the posixly correct behaviour, `getopt` would assume that `-B`
is an option intended for `odkrun`, which would result in an error
because `-B` is not a valid option for `odkrun`. One would need to
insert a `--` ("stop recognising option" signal) to get the expected
behaviour:

  $ odkrun -- make -B