Error out if we can't get the real path of a binding.
When adding a binding to the ODK configuration, if we can't get the real
path of the directory on the host side (meaning that `realpath` returned
NULL), we should fail immediately rather than trying to invoke Docker
with a "NULL" binding.

The only exception is if we can't get the real path because the
specified path does not actually exist. There may be legitimate reasons
for wanting to do that, so we just trust the user in that case.