Remove "seed mode" options.
Remove the --seed option. Seeding mode is now enabled by using "seed" as
the first positional argument, as if it were a normal command.

There is no more need for --gitname and --gitemail options. The runner
will simply try to automatically obtain the Git username and email from
either the environment or the local Git configuration, and pass them to
the container using environment variables if it got them. If we could
not get them, or if the user want to explicitly specify a different
username and/or email, they can use the ---gitname and --gitemail option
as follows:

  odkrun seed --gitname "A. U. Thor" --gitemail "author@example.org" ...

where the options will _not_ be recognised by odkrun, by instead passed
as they are to the `odk.py seed` command.