Add support for run.sh.conf configuration file.
The standard, ODK-generated run.sh script allows to use a small
configuration file, run.sh.conf, that uses the Bourne shell syntax (the
file is just directly sourced at the beginning of the script) and that
is expected to contain variable declarations that alter the default
behaviour of the script.

We add minimal support for such a configuration file. Minimal because we
only attempt to recognise the declarations of the variables we support;
we do not parse the file as if it was a shell script. This means, for
example, that something like this:

  MY_IMAGE=obolibrary/odklite
  ODK_IMAGE=$MY_IMAGE

which would work as expected with run.sh, will not work with odkrunner
as we don't do variable dereferencing.

Other limitations include:

* Binding options in ODK_BINDS (e.g., ':ro' to mount a volume in
  read-only mode) are ignored.
* ODK_USER_ID does not allow to specify an arbitrary user ID; it can
  only be used to specify an ID of 0, which is equivalent to using the
  --root option.