Allow drive letters in binding specs for Windows.
When processing the value of the ODK_BINDS option, we need to account
for the fact that the host directory path on Windows is likely to
contain a ':' character, if the path starts with a drive letter (e.g.
'C:/Users/...' instead of '/Users/...').

While we are at it, we also check that the container directory path is
not empty (that is, the spec does not look like something '/host:').
This does not really matter as Docker will error out on an incomplete
spec anyway, but better to catch that early.