Allow using unqualified image names.
If the name of the Docker image (as specified with the -i option, or in
the run.sh.conf file with the ODK_IMAGE option) is not prefixed with an
organisation qualifier, then we automatically prepend the "obolibrary/"
qualifier.

That is,

  odkrun -i odkfull ...

is now equivalent to

  odkrun -i obolibrary/odkfull ...

This also means it is no longer possible to use unqualified image names
(such as `ubuntu` -- this would become `obolibrary/ubuntu`). But odkrun
is not designed to be used with arbitrary images anyway, only ODK
images.