Allow NOT to override pre-existing values in the ODK configuration.
Change the interface of most functions in the runner.c module so that
they accept a new 'flags' argument.

The only flag defined for now is ODK_NO_OVERWRITE. When that flag is
set, pre-existing values in the ODK configuration object should not be
overwritten.

For example, when adding a new environment variable, if a variable with
the same name already exists and the ODK_NO_OVERWRITE flag is set, the
existing variable is NOT updated (and the call adding the new variable
has no effect).

That behaviour is extended to the function to add bindings. Now, when we
add a binding, we first check whether another binding with the same
host-side path already exists, and if so we update that binding
(previous behaviour would have resulted in the addition of a NEW binding
instead). But if ODK_NO_OVERWRITE is set, then the existing binding is
left alone and we ignore the attempt to overwrite it.

We also add odk_set_image_name and odk_set_image_tag function to set the
name and the tag of the ODK image, respectively. If the ODK_NO_OVERWRITE
flag is set when calling those functions, they only overwrite the
pre-existing values if those values are still the default values (i.e.,
they do not overwrite values that have explicitly been set by a previous
call).

All those changes are intended to allow for finer control of when a
configuration setting can be overwritten.