Load the configuration after parsing the command line.
Now that we can control whether existing configuration settings can be
overwritten or not, we can move the parsing of the run.sh.conf file to
*after* we have parsed the command line options.

Parsing that file before the command line was not really satisfactory as
it meant the file would be parsed even if the program was called with
the --help or --version options, but it was the easiest way to ensure
that command line options would take precedence over the contents of the
configuration file.

Now however, we can parse the command line options first, because when
reading the configuration file all configuration updates are made with
the ODK_NO_OVERWRITE option, ensuring that any setting that has already
been made on the command line will NOT be overriden by the configuration
file.