Load configuration from the environment.
After reading the configuration file (if any), we also look into the
environment for any variable that starts with 'ODK_' or 'OWLAPI_', and
we treat those variables as if they had been found in the configuration
file.

The fact the environment is queried after the configuration file,
combined with the ODK_NO_OVERWRITE flag used when processing the
variables, means that if a variable is found both in the configuration
file and in the environment, the value in the configuration will take
precedence (unless the option has already been set on the command line,
which always takes precedence over everything else). This may be a bit
"atypical" (in many programs, environment variables take precedence over
configuration files), but this mimicks the behaviour of the original
run.sh wrapper script.

This precedence only applies for single-valued options (e.g. ODK_IMAGE,
ODK_SHARE_OAK_CACHE, etc). For multi-valued options (e.g. ODK_BINDS,
ODK_JAVA_OPTS), values found in the environment are *added* to values
found in the configuration file. This means it is not possible to use
the configuration file to "override" a value that exists in the
environment.

To avoid code duplication, we directly reuse the process_line function
that was intended to process the configuration file. This has two small
drawbacks: (1) this is slightly inefficient, in that we have to
systematically make a copy of the environment variable (because
process_line expects to be able to modify the contents of the "line");
(2) it makes any warning caused by a environment variable appear as if
it came from line 0 of run.sh.conf.

Given that the use of environment variables to configure odkrun is
expected to be rare (most users are expected to use either command line
options for one-time configuration or run.sh.conf for permanent
configuration), those drawbacks are deemed acceptable, at least for now.