Silence warning about extra variadic argument.
GCC's -Wall does not like when warnx is called with one more argument
(NULL) than there are format specifiers in the format string, even
though it is necessary here to call the variadic macro (whether that
macro is at all useful is a legitimate question, maybe we should
reconsider that).

Keep the compiler happy by including the value of the ODK_USER_ID option
in the warning message, so that we have a variadic argument to pass.