Add the -m, --java-mem option.
Add an option to allow explicitly setting the maximal amount of memory
that Java applications (mostly ROBOT) are allowed to use.

If used, that option takes precedence over any setting passed through
ODK_JAVA_OPTS, whether from the environment or from a run.sh.conf file.

If not used and no -Xmx option is passed through ODK_JAVA_OPTS, we still
default to using 90% of available memory.

The value of the --java-mem option can be specified in megabytes, in
gigabytes, or as a percentage of the available total memory.